feat: finish move of basic extensions
