+++ packages/pixeloven/cli/src/types.ts
@@ -1,35 +1,12 @@
-import { GluegunCommand, GluegunParameters, GluegunToolbox } from "gluegun";
+import { GluegunParameters, GluegunToolbox } from "gluegun";
 
-interface GetArgListOptions {
-    offset: number;
-    type: "default" | "withOptions";
-}
+export type PixelOvenParameters = GluegunParameters;
 
-export type GetArgListFunction = (
-    cmd: string,
-    parameters: GluegunParameters,
-    options: GetArgListOptions,
-) => string[];
-
-export interface RunResponse {
-    stdout?: Buffer | string;
-    status: number;
-    error?: Error;
-}
-
-export type RunFunction = (args: string[]) => Promise<RunResponse>;
-
-export interface PixelOvenExtensions {
-    getArgList: GetArgListFunction;
-    run: RunFunction;
-}
 export interface PixelOvenOptions {
     [key: string]: string | number | boolean;
 }
 
 export interface PixelOvenToolbox extends GluegunToolbox {
     config: PixelOvenOptions;
-    pixelOven: PixelOvenExtensions;
+    parameters: PixelOvenParameters;
 }
-
-export type PixelOvenCommand = GluegunCommand<PixelOvenToolbox>;
